requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.8
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-quotes==3.3.1
flake8-return==1.1.3
idna==3.3
ipdb==0.13.9
ipython==7.28.0
isort==5.9.3
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
oauthlib==3.1.1
parso==0.8.2
pep8-naming==0.12.1
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.21
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
testfixtures==6.18.3
toml==0.10.2
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
gunicorn==20.0.4
psycopg2-binary==2.8.5
